finishing int->size_t fix, removed (int) cast where now unneeded
git-svn-id: https://svn.code.sf.net/p/teem/code/teem/trunk@4011 3d70eeeb-363e-0410-a505-8a46323a89f2

+++ src/nrrd/write.c
@@ -690,7 +690,7 @@ _nrrdSprintFieldInfo(char **strP, char *prefix,
     } else if (nio->dataFNArr->len > 1) {
       maxl = 0;
       for (ii=0; ii<nio->dataFNArr->len; ii++) {
-        maxl = AIR_MAX(maxl, (int)strlen(nio->dataFN[ii]));
+        maxl = AIR_MAX(maxl, strlen(nio->dataFN[ii]));
       }
       *strP = (char *)calloc(fslen + strlen(NRRD_LIST_FLAG) 
                              + uintStrlen + nio->dataFNArr->len * (maxl + 1),
